move noddy code to pigfarm eric fix too
