+++ karmapi/noddy.py
@@ -20,25 +20,6 @@ from karmapi import pigfarm, toy
 
 class Magic(pigfarm.MagicCarpet):
 
-    def __init__(self, parent=None, data=None):
-
-        super().__init__(parent, axes=[211, 212])
-
-        # hmm.. not sure where this belongs
-        pandas.set_eng_float_format(1, True)
-
-        data = data or toy.distros(
-            trials=1000,
-            groups=['abc', 'cde', 'xyz'])
-
-        self.data = data
-        self.tests = 0
-
-        if data:
-            self.process_data()
-
- 
-
     async def tester(self, sleep=2):
 
         print('NUMBER OF TESTS', self.tests)
@@ -66,48 +47,6 @@ class Magic(pigfarm.MagicCarpet):
         
         await pigfarm.sleep(sleep)
 
-
-    def draw_plot(self):
-
-        group = self.groups[self.group]
-        frame = self.frames[group]
-
-        axes = self.subplots[0]
-        axes.clear()
-
-        # sort columns on mean
-        mean = frame.mean()
-        mean.sort_values(inplace=True)
-        frame = frame.ix[:, mean.index]
-        
-        col_colours = []
-        for label in frame.columns:
-            data = frame[label].copy()
-            data.sort_values(inplace=True)
-            if self.log:
-                patch = axes.semilogy(data.values, label=label)
-            else:
-                patch = axes.plot(data.values, label=label)
-
-            col_colours.append(patch[0].get_color())
-
-        from matplotlib import colors
-        col_colours = [colors.to_rgba(x, 0.2) for x in col_colours]
-
-        self.axes = self.subplots[1]
-        self.axes.clear()
-        self.draw_table(frame, loc='center', title=group, col_colours=col_colours)
-        self.draw()
-        
-    async def run(self):
-
-        await pigfarm.spawn(self.load_data())
-
-        await self.tester(sleep=3.01)
-
-        while True:
-            self.draw_plot()
-            self.group = await self.event.get()
 
 def play(infile):
